requirements.txt
python-telegram-bot==20.7
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
jinja2==3.1.2
aiohttp==3.9.1
python-multipart==0.0.6
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
sqlalchemy-utils==0.38.3
python-jose==3.3.0
passlib==1.7.4
python-magic==0.4.27
pillow==10.2.0
python-dateutil==2.8.2
pydantic==2.5.3
email-validator==2.1.0.post1
httpx==0.25.2
